Add basic kubernetes network support
Rkt uses a netns path that is hardcoded in the CNI plugin. In kubernetes we obtain the PID from the docker client and create a namespace using the PID when provisioning the veth to the calico endpoint.

Also note that this current state uses an outdated version to libcalico. The functions used to assign an IP Address from a pool/subnet are no longer supported.