Get kubernetes Bearer token from env variable
Previously we extracted the Bearer token (used for authenticating
with the kube-apiserver) from the basic auth file. This auth
method is deprecated, so we should instead support the
more generic Service Account tokens.

Both mechanisms use the HTTP "Authentication: Bearer" header,
so all this commit changes is where we get that token from; now
we use the AUTH_TOKEN environment variable.